chore: remove Small component variant
