+++ .eslintrc.js
@@ -38,6 +38,7 @@ module.exports = {
       },
     ],
     "@typescript-eslint/explicit-module-boundary-types": "off",
+    "@typescript-eslint/no-empty-function": "off",
     "@typescript-eslint/prefer-for-of": "error",
     "@typescript-eslint/unified-signatures": "error",
 
@@ -60,6 +61,7 @@ module.exports = {
     "dot-notation": "error",
     "eol-last": "error",
     "eqeqeq": "error",
+    "func-name-matching": "error",
     "func-names": "error",
     "func-style": "error",
     "grouped-accessor-pairs": "error",
@@ -89,7 +91,6 @@ module.exports = {
     "no-div-regex": "error",
     "no-duplicate-imports": "error",
     "no-else-return": "error",
-    "no-empty-function": "error",
     "no-empty": ["error", { allowEmptyCatch: true }],
     "no-eq-null": "error",
     "no-eval": "error",
@@ -124,6 +125,7 @@ module.exports = {
     "no-script-url": "error",
     "no-self-compare": "error",
     "no-sequences": "error",
+    "no-shadow": "error",
     "no-template-curly-in-string": "error",
     "no-throw-literal": "error",
     "no-unmodified-loop-condition": "error",
