+++ src/components/balance.tsx
@@ -1,4 +1,4 @@
-import { formatUsd, translate } from "@galoymoney/client"
+import { formatUsd } from "@galoymoney/client"
 import { SatFormat, SatSymbol, Spinner } from "@galoymoney/react"
 
 import useMyUpdates from "../hooks/use-my-updates"
@@ -10,58 +10,43 @@ const navigateToHome = () => {
   history.push("/")
 }
 
-const MyBalance: React.FC<{ balance: number; layout: "Small" | "Large" }> =
-  function Balance({ balance, layout }) {
-    const { isAuthenticated } = useAuthContext()
-    const { satsToUsd } = useMyUpdates()
+type FCT = React.FC<{ balance: number }>
 
-    if (!isAuthenticated) {
-      return (
-        <div className="balance" onClick={navigateToHome}>
-          {layout === "Large" && (
-            <div className="title">{translate("Current Balance")}</div>
-          )}
-          <div className="value">
-            <div className="primary">
-              <SatSymbol />0
-            </div>
-            <div className="secondary">{formatUsd(0)}</div>
-          </div>
-        </div>
-      )
-    }
+const Balance: FCT = ({ balance }) => {
+  const { isAuthenticated } = useAuthContext()
+  const { satsToUsd } = useMyUpdates()
 
+  if (!isAuthenticated) {
     return (
       <div className="balance" onClick={navigateToHome}>
-        {layout === "Large" && (
-          <div className="title">{translate("Current Balance")}</div>
-        )}
         <div className="value">
-          {Number.isNaN(balance) ? (
-            <Spinner />
-          ) : (
-            <>
-              <div className="primary">
-                <SatFormat amount={balance} />
-              </div>
-              {satsToUsd && (
-                <div className="secondary">&#8776; {formatUsd(satsToUsd(balance))}</div>
-              )}
-            </>
-          )}
+          <div className="primary">
+            <SatSymbol />0
+          </div>
+          <div className="secondary">{formatUsd(0)}</div>
         </div>
       </div>
     )
   }
 
-const Balance = {} as LayoutComponent<{ balance: number }>
-
-Balance.Small = function Balance({ balance }) {
-  return <MyBalance balance={balance} layout="Small" />
-}
-
-Balance.Large = function Balance({ balance }) {
-  return <MyBalance balance={balance} layout="Large" />
+  return (
+    <div className="balance" onClick={navigateToHome}>
+      <div className="value">
+        {Number.isNaN(balance) ? (
+          <Spinner />
+        ) : (
+          <>
+            <div className="primary">
+              <SatFormat amount={balance} />
+            </div>
+            {satsToUsd && (
+              <div className="secondary">&#8776; {formatUsd(satsToUsd(balance))}</div>
+            )}
+          </>
+        )}
+      </div>
+    </div>
+  )
 }
 
 export default Balance
