+++ src/components/login/auth-code.tsx
@@ -7,9 +7,9 @@ import config from "../../store/config"
 import { history, useRequest } from "../../store"
 import { useAuthContext } from "../../store/use-auth-context"
 
-type Props = { phoneNumber: string }
+type FCT = React.FC<{ phoneNumber: string }>
 
-const AuthCode = ({ phoneNumber }: Props) => {
+const AuthCode: FCT = ({ phoneNumber }) => {
   const request = useRequest()
   const [loading, setLoading] = useState(false)
   const [errorMessage, setErrorMessage] = useState("")
