+++ src/components/logout.tsx
@@ -4,7 +4,7 @@ import { Icon } from "@galoymoney/react"
 import { history, useRequest } from "../store"
 import { useAuthContext } from "../store/use-auth-context"
 
-const Logout = () => {
+const Logout: NoPropsFCT = () => {
   const request = useRequest()
   const { setAuthSession } = useAuthContext()
