+++ src/components/root-component.tsx
@@ -7,21 +7,18 @@ import { Spinner } from "@galoymoney/react"
 import { appRoutes, checkRoute, authRoutes, checkAuthRoute } from "../server/routes"
 
 import ErrorFallback from "./error-fallback"
-import { useAppState } from "../store"
 
-type Props = {
-  path: RoutePath
+type FCT = React.FC<{
+  path: RoutePath | AuthRoutePath
   flowData?: KratosFlowData
   [name: string]: unknown
-}
+}>
 
-const RootComponent = ({ path, flowData, ...props }: Props) => {
-  const { layout } = useAppState()
-  const appLayout = layout ?? "Large"
+const RootComponent: FCT = ({ path, flowData, ...props }) => {
   const checkedRoutePath = checkRoute(path)
 
   if (!(checkedRoutePath instanceof Error)) {
-    const Component = appRoutes[checkedRoutePath].component[appLayout]
+    const Component = appRoutes[checkedRoutePath].component
 
     return (
       <Suspense
@@ -32,8 +29,8 @@ const RootComponent = ({ path, flowData, ...props }: Props) => {
         }
       >
         <ErrorBoundary FallbackComponent={ErrorFallback}>
-          <div id="main-container" className={`layout${appLayout}`}>
-            <Component layout={appLayout} {...props} />
+          <div id="main-container">
+            <Component {...props} />
           </div>
         </ErrorBoundary>
       </Suspense>
