+++ src/components/send/send-action.tsx
@@ -3,7 +3,9 @@ import SendIntraLedgerAction from "./send-intra-ledger-action"
 import { SendLnInvoiceAction, SendLnNoAmountInvoiceAction } from "./send-ln-action"
 import SendOnChainAction from "./send-onchain-action"
 
-const SendAction = (props: SendActionProps) => {
+type FCT = React.FC<SendActionProps>
+
+const SendAction: FCT = (props) => {
   let { errorMessage } = props
 
   if (props.satAmount && props.satAmount > props.btcWalletBalance) {
