+++ src/components/send/send-intra-ledger-action.tsx
@@ -3,7 +3,9 @@ import { MouseEvent } from "react"
 
 import SendActionDisplay from "./send-action-display"
 
-const SendIntraLedgerAction = (props: SendIntraLedgerActionProps) => {
+type FCT = React.FC<SendIntraLedgerActionProps>
+
+const SendIntraLedgerAction: FCT = (props) => {
   const [sendPayment, { loading, errorsMessage, data }] =
     useMutation.intraLedgerPaymentSend()
