+++ src/components/send/send-onchain-action.tsx
@@ -3,7 +3,9 @@ import { MouseEvent } from "react"
 
 import SendActionDisplay from "./send-action-display"
 
-const SendOnChainAction = (props: SendOnChainActionProps) => {
+type FCT = React.FC<SendOnChainActionProps>
+
+const SendOnChainAction: FCT = (props) => {
   const [sendPayment, { loading, data, errorsMessage: paymentError }] =
     useMutation.onChainPaymentSend()
