+++ src/components/settings/color-theme.tsx
@@ -6,7 +6,7 @@ import { Icon } from "@galoymoney/react"
 import localStorage from "../../store/local-storage"
 import { ColorTheme, colorThemeLabels, setColorTheme } from "../../store/theme"
 
-const ColorThemeSetting = () => {
+const ColorThemeSetting: NoPropsFCT = () => {
   const [colorTheme, setColorThemeState] = useState<ColorTheme>("DEFAULT")
 
   const handleColorThemeChange = (event: ChangeEvent<HTMLSelectElement>) => {
