+++ src/components/settings/username.tsx
@@ -3,11 +3,9 @@ import { Icon } from "@galoymoney/react"
 
 import useMainQuery from "../../hooks/use-main-query"
 
-type Props = {
-  guestView: boolean
-}
+type FCT = React.FC<{ guestView: boolean }>
 
-const UsernameSetting = ({ guestView }: Props) => {
+const UsernameSetting: FCT = ({ guestView }) => {
   const { username } = useMainQuery()
 
   return (
