+++ src/server/routes.ts
@@ -8,6 +8,7 @@ import Transactions from "../components/pages/transactions"
 
 import Register from "../components/pages/register"
 import config from "../store/config"
+import React from "react"
 
 // Note: The component property is skipped by the serialize function
 // It's only used on the front-end
@@ -47,7 +48,7 @@ const appRoutesDef = {
 }
 
 export type SupportedRoutes = keyof typeof appRoutesDef
-type AppRoutes = Record<RoutePath, { component: LayoutComponent<unknown>; title: string }>
+type AppRoutes = Record<RoutePath, { component: React.FC<unknown>; title: string }>
 
 export const appRoutes: AppRoutes = appRoutesDef as unknown as AppRoutes
 
@@ -77,7 +78,7 @@ export const checkAuthRoute = (path: string): AuthRoutePath | Error => {
 type AuthRoutes = Record<
   AuthRoutePath,
   {
-    component: (props: { flowData?: KratosFlowData }) => JSX.Element
+    component: React.FC<{ flowData?: KratosFlowData }>
     title: string
   }
 >
