+++ src/store/use-auth-context.ts
@@ -2,7 +2,6 @@ import { useContext, createContext } from "react"
 
 export const AuthContext = createContext<AuthContextType>({
   isAuthenticated: false,
-  // eslint-disable-next-line @typescript-eslint/no-empty-function, no-empty-function
   setAuthSession: () => {},
 })
