+++ src/styles/main.css
@@ -594,264 +594,6 @@ button {
   }
 }
 
-.layoutLarge {
-  .header-container {
-    margin-bottom: 2rem;
-    .header {
-      margin-right: 14px;
-      display: flex;
-      justify-content: space-between;
-      align-items: center;
-      .balance {
-        cursor: pointer;
-        .title {
-          text-align: center;
-          margin-bottom: 0.5rem;
-        }
-
-        .value {
-          text-align: center;
-          .primary {
-            font-size: 1.2rem;
-            text-align: center;
-          }
-          .secondary {
-            color: var(--color-secondary);
-            margin-top: 0.2rem;
-            text-align: center;
-          }
-        }
-      }
-      .links {
-        display: flex;
-        justify-content: space-around;
-        font-size: 1.1rem;
-        .menu-icon {
-          cursor: pointer;
-          font-size: 1.25rem;
-          color: var(--color-links);
-        }
-      }
-    }
-  }
-
-  .header-nav {
-    margin: 1rem auto;
-    display: flex;
-    flex-direction: row;
-    a {
-      margin: 1rem auto;
-      border: thin solid var(--color-borders);
-      border-radius: 0.25rem;
-      padding: 0.75rem 1.75rem;
-      background-color: var(--color-light-bg);
-      width: 75%;
-      max-width: 225px;
-      display: flex;
-      align-items: center;
-      justify-content: center;
-    }
-    .active {
-      background-color: var(--color-active-bg);
-      font-weight: bold;
-      color: var(--color-primary);
-    }
-  }
-
-  .home {
-    .header-nav {
-      margin-top: 3rem;
-    }
-  }
-
-  .fee-amount {
-    padding: 0.75rem;
-    border: thin dashed var(--color-borders);
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    margin-bottom: 0.75rem;
-    .label {
-      margin-right: 0.75rem;
-      font-weight: bold;
-    }
-    .fee-usd-amount {
-      text-align: left;
-    }
-  }
-}
-
-.layoutSmall {
-  .header-container {
-    .header {
-      display: flex;
-      justify-content: center;
-      margin: 1rem auto;
-    }
-  }
-
-  .balance {
-    cursor: pointer;
-
-    .value {
-      text-align: center;
-      .primary {
-        font-size: 1.2rem;
-        text-align: center;
-      }
-      .secondary {
-        color: var(--color-secondary);
-        margin-top: 0.2rem;
-        text-align: center;
-      }
-    }
-  }
-
-  .home {
-    .buttons {
-      display: flex;
-      justify-content: center;
-      position: fixed;
-      right: 0;
-      left: 0;
-      margin: 1rem auto;
-      bottom: 0;
-      height: 200px;
-      background-color: var(--color-bg);
-      .button {
-        margin: 1.5rem;
-        padding-left: 1.5rem;
-        padding-right: 1.5rem;
-        height: 20px;
-      }
-    }
-  }
-
-  .send {
-    position: relative;
-    .input-label {
-      text-align: left;
-      width: 100%;
-      margin: 1rem 0 0.5rem 0;
-    }
-
-    .amount-input-container {
-      margin-top: 2rem;
-    }
-  }
-
-  .contact-list-for-sending {
-    margin: 2.5rem 0.5rem;
-    display: flex;
-    flex-wrap: wrap;
-    .spinner {
-      font-size: 1rem;
-      flex: 1;
-      text-align: center;
-    }
-    .contact {
-      display: flex;
-      flex-direction: column;
-      justify-content: center;
-      align-items: center;
-      margin: 0.75rem 1rem;
-      width: 75px;
-      .icon {
-        width: 50px;
-        height: 50px;
-        display: flex;
-        justify-content: center;
-        align-items: center;
-        border: thin solid var(--color-borders);
-        border-radius: 50%;
-        i {
-          margin: 0;
-          font-size: 1.25rem;
-        }
-      }
-      .name {
-        text-align: center;
-        margin-top: 0.25rem;
-      }
-    }
-  }
-
-  .amount-display {
-    .primary {
-      display: none;
-    }
-    .converted {
-      color: var(--color-secondary);
-    }
-  }
-
-  .send-confirm {
-    margin: 1rem auto;
-    .item {
-      display: flex;
-      justify-content: space-between;
-      align-items: center;
-      width: 100%;
-      border-bottom: thin solid var(--color-borders);
-      padding-bottom: 1rem;
-      margin-top: 1.5rem;
-      .content {
-        text-align: right;
-        .amount-display {
-          .primary {
-            display: block;
-          }
-          .converted {
-            font-size: 0.85rem;
-            margin-top: 0.3rem;
-            color: var(--color-secondary);
-          }
-        }
-      }
-    }
-  }
-
-  .footer {
-    position: fixed;
-    height: 50px;
-    bottom: 0;
-    left: 0;
-    right: 0;
-    background-color: var(--color-bg);
-    padding: 1rem;
-    .nav-icons {
-      position: relative;
-      display: flex;
-      .tab {
-        margin: 0 auto;
-        font-size: 2rem;
-      }
-      .tab.disabled {
-        color: var(--color-subtext);
-      }
-    }
-  }
-
-  .fee-amount {
-    margin: 1rem auto;
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    margin-bottom: 2.5rem;
-    .label {
-      margin-right: 0.75rem;
-      font-weight: bold;
-    }
-    .fee-usd-amount {
-      text-align: left;
-    }
-  }
-
-  .action-container {
-    text-align: center;
-  }
-}
-
 .link-copied {
   padding: 1rem;
   margin: 0.5rem auto;
